Add MarshalAs pseduo attribute handling and other refactoring
