+++ src/coreclr/System.Private.CoreLib/System.Private.CoreLib.csproj
@@ -322,11 +322,11 @@
   </ItemGroup>
 
   <Target Name="GenerateEventingFiles"
-         Inputs="@(EventingGenerationScript);@(EventManifestFile)"
-     Outputs="@(EventingSourceFile)"
-     DependsOnTargets="FindPython"
-     BeforeTargets="BeforeCompile">
-    
+          Inputs="@(EventingGenerationScript);@(EventManifestFile)"
+		  Outputs="@(EventingSourceFile)"
+		  DependsOnTargets="FindPython"
+		  BeforeTargets="BeforeCompile">
+
     <Error Condition="'$(PYTHON)' == ''" Text="Unable to locate Python. NativeRuntimeEventSource.CoreCLR.cs cannot be generated without Python installed on the machine. Either install Python in your path or point to it with the PYTHON environment variable." />
     <PropertyGroup>
       <_PythonWarningParameter>-Wall</_PythonWarningParameter>
