+++ src/libraries/System.Reflection.Emit/src/Resources/Strings.resx
@@ -126,9 +126,6 @@
   <data name="InvalidOperation_BadInterfaceNotAbstract" xml:space="preserve">
     <value>Interface must be declared abstract.</value>
   </data>
-  <data name="NotSupported_Signature" xml:space="preserve">
-    <value>The signature {0} is not supported.</value>
-  </data>
   <data name="InvalidOperation_AModuleRequired" xml:space="preserve">
     <value>Assembly needs at least one module defined.</value>
   </data>
@@ -141,11 +138,11 @@
   <data name="NotSupported_DynamicModule" xml:space="preserve">
     <value>The invoked member is not supported in a dynamic module.</value>
   </data>
-  <data name="Argument_InvalidTypeArgument" xml:space="preserve">
+  <data name="Argument_InvalidTypeCodeForTypeArgument" xml:space="preserve">
     <value>The type code may not be used as a type argument of a custom attribute .</value>
   </data>
-  <data name="InvalidOperation_EmptyFieldForCustomAttribute" xml:space="preserve">
-    <value>Custom attribute '{0}' doesn't contain a field named '{1}'.</value>
+  <data name="NotSupported_UnmanagedTypeOnlyForFields" xml:space="preserve">
+    <value> 'UnmanagedType.{0}' named parameter is only valid for fields.</value>
   </data>
   <data name="Argument_InvalidCustomAttributeLength" xml:space="preserve">
     <value>Custom attribute '{0}' data length is only '{1}'.</value>
@@ -164,5 +161,14 @@
   </data>
   <data name="ArgumentOutOfRange_ParamSequence" xml:space="preserve">
     <value>The specified parameter index is not in range.</value>
+  </data>
+  <data name="Argument_InvalidArgumentForAttribute" xml:space="preserve">
+    <value>Invalid constructor argument {0} provided for MarshalAs atttribute.</value>
+  </data>
+  <data name="Argument_InvalidParameterForUnmanagedType" xml:space="preserve">
+    <value>Named parameter {0} is not valid for UnmanagedType.{1} type.</value>
+  </data>
+  <data name="Argument_SizeConstMustBeSpecified" xml:space="preserve">
+    <value>SizeConst parameter must be specified for UnmanagedType.ByValTStr type.</value>
   </data>
 </root>
