+++ src/libraries/System.Reflection.Emit/tests/PersistableAssemblyBuilder/AssemblySaveCustomAttributeTests.cs
@@ -164,6 +164,10 @@ namespace System.Reflection.Emit.Tests
                 {
                     ParameterBuilder pb = meb.DefineParameter(param.Position + 1, param.Attributes, param.Name);
                     paramAttributes.ForEach(pb.SetCustomAttribute);
+                    if (param.ParameterType.Equals(typeof(string)))
+                    {
+                        pb.SetConstant("Hello");
+                    }
                 }
             }
         }
@@ -254,28 +258,29 @@ namespace System.Reflection.Emit.Tests
             {
                 Type dllType = typeof(DllImportAttribute);
                 Type type = typeof(IMultipleMethod);
-                List<CustomAttributeBuilder> typeAttributes = new() { new CustomAttributeBuilder(typeof(ComImportAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
-                                                              new CustomAttributeBuilder(typeof(SuppressUnmanagedCodeSecurityAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
-                                                              new CustomAttributeBuilder(s_guidPair.con, s_guidPair.args)
-                                                            };
-                List<CustomAttributeBuilder> methodAttributes = new() { new CustomAttributeBuilder(typeof(PreserveSigAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
-                                                              new CustomAttributeBuilder(typeof(SuppressUnmanagedCodeSecurityAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
-                                                              new CustomAttributeBuilder(typeof(SpecialNameAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
-                                                              new CustomAttributeBuilder(s_guidPair.con, s_guidPair.args),
-                                                              new CustomAttributeBuilder(typeof(MethodImplAttribute).GetConstructor(new Type[] { typeof(MethodImplOptions) }),
-                                                                                            new object[] { MethodImplOptions.NoInlining | MethodImplOptions.AggressiveOptimization }),
-                                                              new CustomAttributeBuilder(dllType.GetConstructor(new Type[] { typeof(string) }), new object[] { "test.dll" },
-                                                                      new FieldInfo[] { dllType.GetField("CharSet"), dllType.GetField("SetLastError"), dllType.GetField("CallingConvention"), dllType.GetField("BestFitMapping"),
-                                                                          dllType.GetField("ThrowOnUnmappableChar") }, new object[]{ CharSet.Ansi, true, CallingConvention.FastCall, true, false }),
-                                                            };
-                List<CustomAttributeBuilder> parameterAttributes = new() { new CustomAttributeBuilder(typeof(InAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
-                                                              new CustomAttributeBuilder(typeof(OutAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
-                                                              new CustomAttributeBuilder(typeof(OptionalAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
-                                                              new CustomAttributeBuilder(s_guidPair.con, s_guidPair.args),
-                                                            };
+                List<CustomAttributeBuilder> typeAttributes = new() {
+                        new CustomAttributeBuilder(typeof(ComImportAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
+                        new CustomAttributeBuilder(typeof(SuppressUnmanagedCodeSecurityAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
+                        new CustomAttributeBuilder(s_guidPair.con, s_guidPair.args)};
+                List<CustomAttributeBuilder> methodAttributes = new() {
+                        new CustomAttributeBuilder(typeof(PreserveSigAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
+                        new CustomAttributeBuilder(typeof(SuppressUnmanagedCodeSecurityAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
+                        new CustomAttributeBuilder(typeof(SpecialNameAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
+                        new CustomAttributeBuilder(s_guidPair.con, s_guidPair.args),
+                        new CustomAttributeBuilder(typeof(MethodImplAttribute).GetConstructor(new Type[] { typeof(MethodImplOptions) }),
+                                new object[] { MethodImplOptions.NoInlining | MethodImplOptions.AggressiveOptimization }),
+                        new CustomAttributeBuilder(dllType.GetConstructor(new Type[] { typeof(string) }), new object[] { "test.dll" }, new FieldInfo[]
+                              { dllType.GetField("CharSet"), dllType.GetField("SetLastError"), dllType.GetField("CallingConvention"), dllType.GetField("BestFitMapping"),
+                                dllType.GetField("ThrowOnUnmappableChar") }, new object[]{ CharSet.Ansi, true, CallingConvention.FastCall, true, false })};
+                List<CustomAttributeBuilder> parameterAttributes = new() {
+                        new CustomAttributeBuilder(typeof(InAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
+                        new CustomAttributeBuilder(typeof(OutAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
+                        new CustomAttributeBuilder(typeof(OptionalAttribute).GetConstructor(Type.EmptyTypes), new object[] { }),
+                        new CustomAttributeBuilder(s_guidPair.con, s_guidPair.args),
+                        new CustomAttributeBuilder(marshalAsEnumCtor, new object[] { UnmanagedType.CustomMarshaler },
+                                new FieldInfo[] { typeof(MarshalAsAttribute).GetField("MarshalType")}, new object[] { typeof(EmptyTestClass).AssemblyQualifiedName })};
 
-                AssemblyBuilder ab = AssemblyTools.PopulateAssemblyBuilderAndSaveMethod(
-                    PopulateAssemblyName(), null, typeof(string), out MethodInfo saveMethod);
+                AssemblyBuilder ab = AssemblyTools.PopulateAssemblyBuilderAndSaveMethod(PopulateAssemblyName(), null, typeof(string), out MethodInfo saveMethod);
                 TypeBuilder tb = ab.DefineDynamicModule("Module").DefineType(type.FullName, type.Attributes);
                 typeAttributes.ForEach(tb.SetCustomAttribute);
                 DefineMethodsAndSetAttributes(methodAttributes, tb, type.GetMethods(), parameterAttributes);
@@ -364,20 +369,29 @@ namespace System.Reflection.Emit.Tests
                     {
                         IList<CustomAttributeData>  paramAttributes = param.GetCustomAttributesData();
 
-                        Assert.Equal(4, paramAttributes.Count);
+                        Assert.Equal(5, paramAttributes.Count);
                         Assert.True((param.Attributes & ParameterAttributes.In) != 0); // InAttribute
                         Assert.True((param.Attributes & ParameterAttributes.Out) != 0); // OutAttribute
                         Assert.True((param.Attributes & ParameterAttributes.Optional) != 0); // OptionalAttribute
+                        Assert.True((param.Attributes & ParameterAttributes.HasFieldMarshal) != 0); // MarshalAsAttribute
+
+                        if (param.ParameterType.Equals(typeof(string)))
+                        {
+                            Assert.Equal("Hello", param.DefaultValue);
+                        }
 
                         for (int i = 0; i < paramAttributes.Count; i++)
                         {
                             switch (paramAttributes[i].AttributeType.Name)
                             {
                                 case "InAttribute":
-                                    break;
                                 case "OutAttribute":
-                                    break;
                                 case "OptionalAttribute":
+                                    break;
+                                case "MarshalAsAttribute":
+                                    Assert.Equal(UnmanagedType.CustomMarshaler, (UnmanagedType)paramAttributes[i].ConstructorArguments[0].Value);
+                                    Assert.Equal(typeof(EmptyTestClass).AssemblyQualifiedName,
+                                        paramAttributes[i].NamedArguments.First(na => na.MemberName == "MarshalType").TypedValue.Value);
                                     break;
                                 case "GuidAttribute":
                                     Assert.Equal(s_guidPair.args[0], paramAttributes[i].ConstructorArguments[0].Value);
